refactor: adjust malt imports
